Updated Code
MC3_BACKEND and MC3_RUNFILE split the functions and run file into two separate scripts, hiding the code required for backend. The structure required for the chemistry and thickness files is attached in the Excel sample files. RUNFILE allows for the code to be run with just a few directory name changes.